Trait vortex_expr::VortexExpr

pub trait VortexExpr:
    Debug
    + Send
    + Sync
    + PartialEq<dyn Any> {
    // Required methods
    fn as_any(&self) -> &dyn Any;
    fn evaluate(&self, batch: &Array) -> VortexResult<Array>;

    // Provided methods
    fn collect_references<'a>(&'a self, _references: &mut HashSet<&'a Field>) { ... }
    fn references(&self) -> HashSet<&Field> { ... }
}

Represents logical operation on Arrays

Required Methods

fn as_any(&self) -> &dyn Any

Convert expression reference to reference of Any type

fn evaluate(&self, batch: &Array) -> VortexResult<Array>

Compute result of expression on given batch producing a new batch

Provided Methods

fn collect_references<'a>(&'a self, _references: &mut HashSet<&'a Field>)

Accumulate all field references from this expression and its children in the provided set

fn references(&self) -> HashSet<&Field>

Accumulate all field references from this expression and its children in a new set